Add TD breakpoints to spiceSinOp and spiceExpOp
I was getting very minor comparison failures in a few tests.  I
discovered, eventually, that these failures were in rad test cases that
relied on the spice_sin operator.  I then realized that I had not
implemented a getBreakPoints function in that operator, in the new
expression library.

This had seemeed reasonable at the time, as the N_DEV_SourceData.C
SinData class doesn't have a breakpoints function either, and also
sinwaves don't have discontinuities.

What I belatedly realized today was that although a pure sinewave
doesn't have any breakpoints, the spice_sin implementation allows a
delay parameter, TD, after which the sinewave starts.  So there is a
discontinuity at this delay point.  The same is true for the spice_exp
source, except that it has two of these.

The old expression library breakpoint doesn't depend on
operator-specific breakpoint functions.  It uses the same breakpoint
detection algorithm for all kinds of sources.  So, it was detecting
these TD-based breakpoints in spice_sin and spice_exp.  But for new
expression to detect them I had to explicitly implement them.  That is
done now in this commit.