First (insufficient) cut at breakpoints for stp op
The STP function is one that needs to provide a breakpoint.  Up until
now, I had not implemented a getBreakPoints function in it however.

The STP function is a step function :  It is invoked as stp(x), and
returns 1 if x > 0, and zero otherwise.  So

I have now implemented a first pass of the stpOp::getBreakPoints function.
It is not completely sufficient, however.  This function can fail if
the stpOp is being called from inside of a .func.  For similar reasons,
the sdt and ddt operators (in their current implementation in new
expression) can also fail when invoked in a .func.

I have been thinking about how to fix this (for stp, sdt, and ddt) but
have not implemented it yet.

The reasons functions like this can fail inside of a .func is related to
the (intentional) design of the new expression library.  In the new
expression library (unlike the old one), .funcs are not handled with
string substitution.  They are handled by attaching pointers.

But, .funcs have function arguments, which have to be passed in
properly, and (in the case of ddt and sdt) they also have internal state
that is tied to those function arguments.   So any operations on these
operators that involve those function arguments, must be done with the
proper arguments substituted in.   Same goes with any internal state.

This all gets handled naturally in the way that I wrote functions like
"val", which traverse the entire AST, and evaluate everything in the
correct order.  So, if function arguments are themselves expressions,
they are naturally evaluated as one goes up the tree.

The code I wrote in the stpOp requires that the argument to it (stp(x),
where x is the argument) be evaluated correctly.  That is because the BP
code is based on the formula  :

        breakpoint_time = current_time - F(x)/F'(x)

Where "x" could be an expression such as x=time-delay

So, F(x) and F'(x) have to be correct, "x" must be evaluated properly.
For that to be possible, it has to be done during an evaluation, or
something like an evaluation.  ie, a full traversal so that if "x" is a
.func argument, it is set up correctly.

But at the moment, the "stpOp::getBreakPoints"  function is a
stand-alone function, that is not called during a traversal.  That means
that if it is part of a call to a .func, it will not have the correct
"x".

Similar problems happen with sdt and ddt; when values are pushed into
the history, they have to be done in the context of properly resolved
function arguments.  Otherwise they can be wrong.

So, for the moment, breakpointing in stpOp and time-based calculus in
sdt and ddt are busted for the .func use case.  I have been thinking
about how to fix this, and hope to have a solution soon.