Finish breakpoints for stp and various compare ops
In my previous commit I had an initial implementation of breakpoints for
stp.  I mistakenly referred to these as non-deterministic breakpoints,
but that isn't really true.  In Stp, if the argument depends linearly on
time, then it is a completely deterministic computation.

That previous checkin was incompletely because (as I detailed in my
previous comments), it didn't always work if stp was inside of a .func,
as it depended on the function arguments being set.  In the
"getBreakPoints" function, they were not set correclty, and there was no
easy way to set them correctly.

So, to get around this, I now compute breakpoints during the "val"
function and store them.  I've done all the work to make the function
arguments be correct in the "val" function, so this works.  I also added
a bunch of code to make sure i wasn't returning the same breakpoint
hundreds of times.  (this part isn't perfect, but it is better than
nothing).

Finally, I also generalized this code so that it could be used by 6 of
the comparison operators.  Those 6 comparison operators are things like
greater than, equals, less than, etc.     So, now these operators also
support breakpoints when you have something like "if(time > 2ms,5,4)" or
something like that.

For all of these breakpoints calculations, I have basically followed
what the old expression library did, which is to compute a single step
Newton solve.  Being single step, of course, it assumes that the
dependence on time is linear.  If someone specifies an SDT with an
argument that depends nonlinearly on time, then this will break.
Currently, the breakpoint code does not attempt to check if there is
nonlinear dependence.