Small lexer fix to handle params starting with "E"
I had a minor goof in the lexer, where in several of the regex's that I
use to lex numbers, I used a star special character (meaning 0 or more)
rather than a plus (meaning one or more).

Spefically, for numbers that used scientific notation with an exponent,
I needed to have {DIGIT}+ rather than {DIGIT}* immediately preceding the
[eE].  Since I used a star, that meant that a parameter named "E1" was
interpretted as a number, which was not my intent.

I discovered this because my adding breakpointing to STP as well as
comparison operators were intended (among other tests) to fix the
BUG_174/nlrcs10.cir circuit, and several other tests that are derived
from it.  But, much to my surprise, it did not fix these tests at all.
They were getting comparison failures.

The reason they were getting comparison failures is that they had two
parameteres "E1" and "E2" that the lexer was parsing as numbers rather
than as params.  So, they always evaluated to 0.  Not surprisingly this
resulted in the answer being wrong.

It was my hope that this fix would magically fix a bunch of other tests.
But alas, like most of my fixes these days, it fixed the 2 or so tests
that I knew about.

At this stage, the new expression library is down to 38 test failures
when run against taglist="+nightly+serial?klu?rad?qaspr".  As I have not
yet done any fixes related to error handling, this is getting into the
ballpark of about 20 *real* failures, which is pretty good at this
stage.  I suspect that a bunch of them are going to be very small
one-off, single test failure issues.