Rewrite of breakpoint function for comparison ops
Most of my recent commits have all been about this topic.  I thought I
mostly had this problem fixed, but I realized yet again that I wasn't
quite right.

What led me to this realization was that with all of my recent fixes,
Xyce went from passing the ABM_BREAK/break.cir test to failing it,
which was a surprise.  The break.cir test has several Bsrc's in it,
each of which produce different breakpoints.

Only the 3rd of those sources pertained to my recent work.  The others
were handled in other ways.  And, with my changes leading up to this
commit, the break.cir circuit went from passing to getting stuck with
really tiny time steps.

The source of this was the fact that the third Bsrc uses a comparison
operator, but the inputs to that operator are a nonlinear function of
time.  I had assumed (incorrectly) that the only use cases of interest
would be linear.  So, I hardwired my breakpoint function to only do a
single Newton iteration to find the breakpoint.  For the break.cir test
case, this mean I had an unconverged answer, and this caused all sorts
of havoc.

So I have rewritten this function to be a true Newton loop.  The old
expression library had also done a Newton loop, but it had been hard for
me to figure out what the heck it was doing, so I missed this.  The new
expression library now passes the break.cir circuit again, getting almost
exactly the same number of time steps as the old expression library.
Also, I think this function is much more readable than its counterpart
in the old library.

More importantly, I've forced this function to use the breakpoint
tolerance, bpTol from the time integrator as the solver tolerance
for computing these breakpoints.  This is a tigher tolerance (I
think) than the old code was using.  The old code had a problem
with computing breakpoints that were closer together than an
arbitrary epsilon.  That was mostly fixed in the new library by
my NOT using Newton's method for most breakpoints.  But for the
case where I have to use Newton's method (comparison ops), I
am hoping my tighter tolerances will mitigate that old issue.

With this commit, the new expression library now only fails
37 regression tests, when run against the test suite with
the following tags (I realize there are some tests not
covered by this list):

--taglist="+nightly+serial?klu?rad?qaspr "

All Test runs complete - Warning 18 Tests FAILED
                                 19 Tests EXITED WITH ERROR