Removes unnecessary Epetra/Ifpack interfaces and adds abstract Operator class
One of the missing linear algebra abstractions in Xyce that makes migrating
from Epetra to Tpetra difficult is the absence of an abstract operator.  The
operator became necessary during the development of the HB analysis capability.
At that time, the "MatrixFreeEpetraOperator" class was created and integrated
into the Nonlinear solver package to bridge the need for applying the Jacobian
matrix to a vector instead of assembling it.  This class was written directly
to the Epetra_Operator class.  At the very least, the MatrixFreeEpetraOperator
class needed to be moved into the Linear Algebra package.  However, the
creation and integration of an abstract operator is ultimately necessary and
that is the focus on this commit.

This commit performs the following changes:

- Creates Linear::Operator abstract class.
- Moves MatrixFreeEpetraOperator to Linear Algebra package and uses that class
within the LinearProblem class to create the Epetra_Operator.
- Changes nonlinear solver class to inherit from the new Linear::Operator
class, where the applyJacobian() method is applied for the operator apply method.
The nonlinear solver object is then used to create the HB linear problem object
for the linear system.
- Found unused infrastructure for creating Ifpack preconditioners in the
nonlinear solver package, which has been removed.
- Found unused Epetra headers in the nonlinear solver package that have been
removed.
- Found unused code for computing gradients in the nonlinear solver "shared"
interface, which has been removed.

So, this is part refactor and part cleanup of things found during the refactor.
Epetra and Ifpack are no longer directly called in the nonlinear solver package.
Dead code has been removed from the nonlinear solver package.

This commit relates to issue #25