Add getUnresolvedFunctions API function
I have been debugging another regression test that has been failing with
new regression.

That test case is BUG_547_SON/mb_orig.cir.

The reason that it is failing is that the new expression library is
confused about a current variable.  It determines that this current
variable is a lead current when it is not.  It is used in a Bsrc
expression, and lead currents are not allowed there.   So, this is an
error in the new expression library I need to fix.

Unfortunately, that wasn't the error message emitted from Xyce when I
first tried to run it.  It complained that it couldn't resolve a
function call, to a .FUNC called DM_AC.  But, when I dug into the code,
it was resolving this .FUNC just fine.

What I discovered later is that the DM_AC error message was a bug.  In
the N_IO_DeviceBlock::setParameterValues function, it attempts to
resolve the V={} parameter, by calling the
circuitContext_.resolveParmaeter function.  If a "false" comes back, it
then emits an error message.  To construct that error message, it
queries the relevant expression for unresolved symbols.

In the place where I should have had it calling "getUnresolvedFunctions"
it was instead calling "getFuncNames".  So, it wasn't checking if a func
was resolved or not, it was gathering its name unconditionally and
adding it to the error message.

And, to make matters worse, the function I should have been calling
"N_UTL_Expression::getUnresolvedFunctions", didn't exist yet.

So, this commit implements the function, and calls it correctly.

Now, BUG_547_SON/mb_orig.cir still fails, but the error message is
closer to the right one.  Now to fix that issue ...