Cleanup, adding more macros to the lexer
I have a lot of rules at the moment for numbers with suffixes and/or
units.  Part of why there are so many rules is that my regex isn't yet
good enough to have figured out how to combine some of them.  That can
come later however, as an optimization.

For the moment, as long as I have a lot of rules that are kind of
similar, it makes sense to make more use of macros.  So, if I've left
out a particular unit (like Henry, which I discovered today), I now only
have to change the macro, and I don't have to edit every single rule
that might need the change.