Fix to current variable logic for subckt currents
When a current variable is used in a Bsrc expression, it must be a
solution variable current and not a lead current.

The new expression parser has logic to determine which type of current
something is.  If it is a solution variable current, it allocates the
currentOp node, and if it is a lead current it allocates the
leadCurrentOp node.  There are other details as well, that are handled
differently.

Also, if the current token is something like "IE", then it is clearly a
lead current as well.

But for the case where the token is just "I", then the only way to tell
is by looking at the argument, which is the name of a device.  Device
types are tied to their first character, so if the name of a device is
Vxxx, then it is an independent voltage source, and you know that it is
a solution current.

I forgot to account for devices inside of subcircuits.  So, when my code
logic encountered a device starting with "X", it mistakenly decided it
was a lead current and allocated the wrong class.  I have added a tiny
bit of extra logic to handle the subcircuit case.

With this change, three more circuits, including BUG_547_SON/mb_orig.cir
now pass.