Small fix to error reporting for expr-dep caps
The test CAPACITOR/ErrorMessageTest.cir has been failing with new
expression.    This particular netlist has a LOT of intentional errors
in it, and the script checks to see that the correct error messages are
emitted for each one.

When run using the development branch (and the old expression library)
resulted in 11 errors reported.

Prior to this commit;  running with the fork with the new expression
library resulted in 10 errors being reported.  So, all but one were
reported correctly.

The one that was missing had to do with ddt not being allowed in
expressions used by capacitors.  I had commented that bit of code out
earlier in my work on the new expression library.  It didn't need to be
commented out (I think I had it commented out b/c I had not yet
implemented the getNumDdt function), and this code was the code
that would have emitted the missing error message.

This commit restores that code (i.e. uncomments it).

This doesn't *quite* fix the test, but the remaining problem is one that
I think doesn't matter.   The only differrence remaining is in one of
the error messages:

 '1e-06*I(R9)'. vs '1e-6*I(R9)'

other that that, it is fixed.