Improve Handling of WHEN Time for DERIV-WHEN Measure
This commit fixes the handling of the WHEN time for the DERIV-WHEN
measure, for AC, DC and NOISE measure modes, if either FROM or TO are
specified.  The DERIV-WHEN measure is now only satisfied if the
interpolated WHEN time is within the specified measurement window.
This commit also further ensures that the AT time, for a DERIV-AT
measure, must be within the FROM-TO window for a successful measure.

Additional work on this issue is still needed for TRAN mode measures.

SON Bug 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304