Correct global param op builder logic
The initial parsing of .print line outputs turns the output list into
a list of "tagged params" (N_UTL_Param.h's Param class) that somewhat
cryptically encode things like "V(A)", "I(M5)", "N(M1:vds)" and so
forth as a sequence of params.  Simple outputs like global params are
a single element in this list, as are complex expressions (which are
decoded elsewhere).

Later, this ParamList is passed off to an "Op builder manager" that
tries to construct "Op" objects that retrieve the desired item out of
the solution vector, state or store vectors, or the appropriate global
parameter storage.  It does so by cycling through an entire container
of op builders, each of which is expected either to return 0 if it
doesn't recognize something or a pointer to an op if it does recognize
it.

The op builder for global parameters was inappropriately matching its
name even if that name happened to be the first element of a sequence
of params that represented an accessor.

So:
   .global_param v=1

   ...
   .print dc V(a)

was being mishandled, solely because the global param op builder saw
the first element of the param list "V" and that it was the name of a
global param, without noticing that this V was the first element of a
pair of items that represented V(A).

This commit fixes the logic of the global param op builder.  The
conditional added here is basically the negative of the one in the
"parameterNameAndArgs" function of N_IO_OpBuilders.C, which is meant
to recognize accessors like V(a), VI(B), I(M5), ID(M5), and the other
special items like DNI(), DNO(), Y(), S(), etc.  Only if the front
element of the param list passed to this op builder is NOT the first
of such an accessor will the global param op builder now even look to
see if the item matches a global param name.

This fixes bug 812 (SON) and also addresses a question posed last week
on the Google Group regarding a .global_param named "ID" that
interfered with printing the drain current of a MOSFET.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=812