Initial commit of graph interface and abstraction
Another abstraction that has been missing from the Xyce linear algebra
layers is the notion of a graph class.  This means that when a class
needs information from a graph, the Epetra_CrsMatrix was accessed
through the epetraObj() method and then the Epetra_CrsGraph was accessed
through the Epetra matrix.

At this time the Linear::Graph class just encapsulates an Epetra_CrsGraph.
The constructor takes an Epetra_CrsGraph and methods allow access to that
object through an epetraObj method.  Once this class is inserted at an
intermediary for the CrsGraph class, it will be expanded to include
necessary methods to support interfacing to graphs.

This commit addresses issue #25