Update fixed and known defects tables
Eric fixed bug 807 for release 7.1, but the 7.1 release notes still
had it as a known defect.

It's too late to fix that for 7.1, but I've taken care of it for the
7.2 release notes.

Also, the known defect issue 970 was a little too broadly stated, and
Eric had done a bunch of work scoping out what the real problem was.
I'm updating the known defect to reflect what he worked out.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=807
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=970