Remove 807 from fixed defects
As it turns out, the 7.2 known defects table was copied from the 7.1
known defects *before* 7.1's release notes were updated, and so it
incorrectly showed 807 as a known defect.  In my last commit I moved
it to fixed with a note that it had been done for 7.1 and left out of
the release notes, but in fact it was already in there.

So it just needs removing from the 7.2 version.