Integrates new Graph class into block linear algebra and analysis types
The block linear algebra and analysis types use the graph from the time
domain linear algebra, combined with the block pattern, to generate the
block graph.  Nothing about the graph is used, it is just passed
from the parallel distribution manager to block linear algebra classes.
So, this makes the best first test of the new Graph class.

Xyce regression passes with these simple changes that extend the
block linear algebra to accept the Linear::Graph and use those extensions
in AC, MOR, and NOISE.

This commit relates to issue #25