Fix for bug 1032 SON, handling complex .print ops
The new expression library supports complex numbers.  That means that
the outputs on the .print line now have a LOT more options.  However,
there were some details that needed to still be worked out.  Among other
things, if you have .print AC {V(A)}, it only prints the real part of
V(A), even though both the real and imaginary components are available
now.

My main change was to expand the ExpressionOp class to also have
ExpressionRealOp and ExpressionImaginaryOp classes. This was necessary
to support automatic output of Re() and Im() complex components.

I also had to add new API functions to the ExpressionData class, to
support std::complex<double>, so that this class could optionally
return a complex number from the new expression library.

The one bit of weirdness is that currently, this capaability will
unconditionally output Re() and Im() for all expressions (when running
with AC, that is).  So, even if the print line says:

.print AC {VI(B)}

I get Re({VI(B)}) Im({VI(B)}) in the output file.  And, for the VI
operator it is extra weird (but maybe ok) because VI is technically a
"real" number and not imaginary.  So, if VI(B)=1.0, then the output is

Re({VI(B)}) Im({VI(B)})
  1.0        0.0

Which seems kind of counter intuitive.  But the information is there.
With this commit, I can now have things like:

.print AC {V(B)*(1.0e+2.0J)}

and get completely correct Re() and Im() components in the output file
for the expressions.