Connecting error handling to Xyce error mgr
flex and bison have a nice infastructure for error handling.  I have
mostly treated it as an afterthought so far, as I mostly cared about
trapping errors in a debugging context rather than from the user point
of view.  However, the expression library is pretty well under control
at this point, and a lot of the remaining test failures are related to
error messages, and also a few cases where the expression library is fed
a broken expression, but doesn't catch the error.

So, I am working on that now.  One of the first steps is to do more to
integrate the flex/bison error handling with the Xyce error handling.
In flex/bison, it is your responsibility to provide error functions
which overide the flex/bison error functions.  That way you can
customize what the error reporting functions do.  As the original
development of the expression library was 100% stand alone, and not
integrated into Xyce, I put this off, and just had the overide error
functions emit messages to std::cout.  That was fine for a while.

With this commit, for the first time I am having the error functions
fall the Xyce::Report function.  Making this work wasn't very hard in
the new-expression build of Xyce, but required that a lot more ERH and
UTL object files be compiled into my unit tests.  This commit, which is
he first step of many, takes care of that.    With this commit, I have
changed the error function at the bottom of ExpressionParser.yxx so that
it calls Xyce::Report instead of std::cout, and I updated the two main
Makefiles so that enough UTL and ERH files are included in the unit
tests for this to work.

I have NOT yet updated the yyerror function in the lexer.  That
probagbly comes next.  I have also not fully fleshed out the call to
Xyce::Report.   I may change it to be a warning instead of an error, to
let the calling code do some subsequent handling.  Or I may not.