Fix for bug 850 SON test.  Also, work on err msgs
This commit does more to connect the error trapping of the lexer and
parser to the error message system used by Xyce.

It also adds new code to the newExpression::attachFunctionNode function,
to throw errors at this stage (via Xyce::Report) when a function is
called with the wrong number of arguments.    I had trapped for this
before, but in the funcOp class in ast.h, and the error trap there had
not used Xyce::Report.  So, the error message came too late, and it also
was essentially just a print statement.

So, this commit gets the new expression library to a point where it
matches the old expression library behavior for bug 850.  This is good.
But, the test doesn't quite pass b/c the error messages don't 100%
match.  I think the differences aren't worth fixing at this point, so I
plan to update the test.  For the record, the differences at this point
are of this nature:

<  UDFA(MU,SIGMA) in expression UDFA(10M,1U,23)+V(2)
---
>  UDFA(MU,SIGMA) in expression UDFA(0.01, 1e-06, 23)+V(2)

In other words, saying the same thing, but in the new expression library
it preserves the suffixes from the netlist, and in the old one
they are expanded to their non-suffixed numbers.  I think the new
expression library output is more useful, as it makes it easier to grep
for successfully.

Also, leaving the suffixes un-expanded is more natural in the new
expression library.  I'd have to do a bunch of extra work to do the
expansion, and I don't see the point of that.

There are still a few places where new expression fails to catch an
error, and I plan to work on that next.