Make buildxyceplugin check for more errors
buildxyceplugin requires a list of Verilog-A input files followed by a
directory name.  It has an optional argument "-o <libname>" that will
be used as the library name if given.  If not given, the library name
is called "Xyce_Plugin_" with the module name of the last file in the
list appended.

Until this commit, a command line error where a user forgot to give a
library name would cause the first Verilog-A file name to be used as
the plugin name (with ".so" appended), and if there were no further
verilog-a files on the command line, wind up producing a null plugin
with no error messages or warnings.

This commit checks that the list of verilog-a files is not empty, and
if it is exits with a usage display rather than slogging on and doing
something useless.

This is in response to a query on the google group yesterday.