Make buildxyceplugin more verbose
At one time, I thought that having buildxyceplugin shut up most of the
time was a good thing.  Its silence can lead a user to overlook some
problems.

Yesterday a user reported that his plugin wasn't working properly, and
it turned out that the reason was that he'd left off the plugin name
from the "-o" option, and so the first (and only) verilog-A file name
he provided was used as the base name for the plugin, and then there
were no files to compile.  The script gleefully pressed on and made a
plugin library with nothing in it.  That type of error is no longer
possible, because the script now checks for a null list of things to
compile.

However, if the user had given more than one file to compile and made
the same error (omitting the plugin name while giving the "-o"
option), the first file in the list would still be used as the plugin
name, and because the list is not empty, the script would produce a
plugin just fine, but missing the first file.  I can see no way to
guard against this sort of error without a lot more smarts in the
script.

Instead, I'm making the script announce what it is doing in more
detail.

This commit makes the script announce that it is building a plugin
from a list of Verilog-A files and print the list of files.  The user
can then look at that list to see if something's missing.

It also announces that it is producing an output plugin and, if the
user specifed -o, what the plugin name will be.  If the user doesn't
specify a plugin name, the plugin name is auto generated later in the
script, so all we say is that it will be automatically generated based
on the name of the last module.

This is by no means bulletproofing the script, but it provides a user
with more information to diagnose what it's doing and how that differs
from what they *intended* it to do.