Correct the -plugin help text
When "-plugin" was added as a command line option, its help was never
made correct.  It simply stated that the command line option was
"-plugin," whereas the correct usage is that "-plugin" must be
followed by a comma-separated list of plugin library paths.

The fact that it can be a comma-separated list was documented exactly
nowhere until right now.  I'm fixing it in the Xyce/ADMS users' guide,
too.