Integrates new Linear::Graph class into PDS and LAS layers
The Linear::Graph class, which is essentially a container for the
Epetra_CrsGraph class, is being integrated in the Parallel::Manager
and Linear Algebra classes (Builder, Matrix, and block linear algebra).

The graph registered with and returned from the Parallel::Manager is
now a Linear::Graph object.  These objects are often just retrieved
from the Parallel::Manager and provided to the appropriate Linear::Builder
for each analysis type.  There is no need to know anything more
specific about the graph in the analyses.

This step will enable the understanding of which graph functionalities
are necessary for the builders.  Furthermore, the current needs of
the graph implementation may not be compatible with the Tpetra design.
This refactor will make such requirements more clear.

This addresses GitLab issue #25