Fixes const correctness in interface to graph objects
The Parallel::Manager should return a const pointer to the graph object.
Historically, it has always returned a non-const pointer, but that might
encourage bad behavior.

This final commit in the new Linear::Graph integration fixes the const
behavior of interfaces that use the graph object.  It also removes some
remaining Epetra headers that are not necessary.

Addresses GitLab issue #25