Completely remove Xygra
The Xygra device was intended as a one-off coupling mechanism to
Alegra ("Xyce/Alegra == Xygra") for a particular use case.  It grew a
little to be a slightly more general tool for coupling to EM codes,
but was too limited and kludgy to be everything Alegra needed for
other use cases.

The General External Device was written to replace Xygra in Alegra,
and was intended to be general enough that other use cases wouldn't
have to be twisted to fit Xygra's limitations.

At this time, all known users of Xygra have migrated to GenExt, and
Xygra has been marked as deprecated for over a year.  It's time to let
it go and simplify the code base.

This commit removes all traces of Xygra from Xyce, including all of
its test harnesses.
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=996