Remove "legacy" Xyce/ADMS that used Sacado
The new non-sacado Xyce/ADMS has had a few releases of testing and it is
now safe to remove the old one, which is no longer in use.

While I'm at it, I'm removing the "obsolete" examples, which include
bsim6.1.0 and vbic 1.2.  These have not been in Xyce for some time,
and there is no value to retaining their source code in current
checkouts.  They live forever in Git and can be retrieved if needed.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/31