Fix for w!=2.0u, and better error handling.
I have generally been working on error handling in the new expression
library.  Also, as of today I got access to a GF 45nm circuit.
(translated via XDM)

Initially, the circuit would not run in Xyce with the new expression
library.  This was for two reasons.

Reason (1) was that in the PDK a lot of real valued numbers were being
speficied as 3., rather than 3.0.  I had not made the regex for numbers
in the flex input file quite flexble enough to handle this.  That was
really easy to fix.  But kind of surprising that nowhere in the Xyce
regression suite does this (apparently).

Reason(2) was that I had incorrectly allowed the character "!" be part
of a parameter name.  That meant that when an expression said something
like w!=2.0, the parser thought it was saying that a parameter named
"w!" was equal to 2.0, instead of a test to see if the parameter named
"w" was NOT equal to 2.0. I have fixed this mistake.

This commit also cleans up some goofiness in the ExpressionLexer
constructor.  I was passing in one argument that was useless, and
unused.   I also added an argument to it so that the expression string
is known to the lexer (and not just the parser).  That makes errors that
come from the lexer a bit more informative.  When I got errors related
to (1), above, it would just say "Unrecognized character . ", which
didn't tell me much that was useful.