Support $bound_step in Xyce/ADMS
$bound_step(expression) is a Verilog-A "system task" that is supposed
to tell the time integrator not to take time steps larger than
expression.  It is useful in analog behavioral models that are
effectively independent sources.

A user on the Google group was trying to simulate a voltage controlled
oscillator using a behavioral model --- the instantaneous frequency of
the output sine wave was dependent on the input voltage. Xyce
gleefully takes huge time steps and misses all the behavior of the
device.

The original Verilog-A the user had been basing his model on had used
$bound_step to avoid precisely this problem.

Xyce has always supported the capability of a device dictating a
maximum time step. $bound_step simply exposes that mechanism in
Verilog-A.

$bound_step is what ADMS calls a "callfunction" and ADMS helpfully
enumerates all callfunctions used in a module in a single
datastructure.  Supporting this feature simply requires that we look
in that datastructure to see if $bound_step is called at all.  If so,
we overload the instance base-class function getMaxTimeStepSize to
return the value of a private instance-class variable "maxTimeStep_".
We use the same conditional to declare and initialize that variable.

After that, anywhere that $bound_step is called, we simply emit an
assignment intto maxTimeStep_.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/32