Bump version number in Xyce/ADMS
Xyce/ADMS emits a block of header comments with the date and time that
the C++ files were generated, and also a comment that indicates what
version of Xyce/ADMS was used to generate it.

I forgot to bump this number in 7.0 and 7.1, so all the C++ code we
have that as generated since 6.12 has claimed it was built with a
back-end intended for Xyce 6.12.

There have been enough changes since 6.12 that it really shouldn't lie
like that.  Since it's just headers in the C++ files I'm not too
worried, but it shouldn't be wrong.