Remove stripping of surrounding curly braces
In Xyce, expressions are  usually specified with curly braces around
them.  They can also be specified with single quotes, but these are
auto-converted early in the parsing to curly braces.

Anyway, for some reason early in this integration process I hacked in a
little bit of code to remove curly braces from the expression string,
but only if they were the beginning and ending characters.

But I also apparently have a rule in the parser grammar, which treats
curly braces exactly the same as parenthesis.  (not all parenthesis,
just parens that are free of other grammar rules).  Since that rule is
there, there isn't any reason to manually remove the curly braces before
flex/bison receives the expression string, I think.

Also, as it turns out, retaining the curly braces allows the grammar to
be a little better at catching some kinds of errors, without my having
to create new rules for catching them.   For example, as
it is currently written, the new expression parser will not flag this as
an error:  I(VM),V(2)*V(3)  but it will flag this as an error
{I(VM,V(2)*V(3)}.     It *should* flag both, and I need to think about
that a bit.  But at the moment, leaving the curly braces in place (as
they were there anyway) seems to catch some errors that were not
being previously caught.