Add yet more rules to the lexer for numbers
I recently modified the rules the produce the TOK_NUM token (for real
numbers), to handle things like 1.  (a number with a decimal point
followed by nothing)

This was necessary, but not sufficient.  I also needed to handle things
like .1  (a number preceded by a decmial point and nothing else).
I also neeed to handle both of the above for the exponential form (such
as 1.e-7 and .1e-7).

This was revealed when I restored curly braces to the expression
string passed to flex/bison (per my last commit).  Restoring those
curly braces made the flex/bison error checking more reliable.

The specific regression test that failed once I restored curly
braces was Certification_Tests/BUG_80_SON/Bad_PWL_Source.  That
circuit is one of the few in Xyce that has numbers like above, including
this expression: {4.*pi*1.e-7}.

This test (per the name) is actually an error message test, and the
error message being tested is for a malformed PWL source.  So, prior to
this work, the new expression library was "passing" this test, but
probably silently failing to parse expressions like {4.*pi*1.e-7}.
The fact that the new expression library was silently failing to parse
{4.*pi*1.e-7} wouldn't have affected the test.

But now, with curly braces restored it was explicitly failing to parse,
no longer silent.

With this commit, I have added a bunch of new rules for TOK_NUM in the
lexer input file.  Hopefully this will be the last time, knock on wood.
I now have different rules for exponential numbers and also for
different variants of units, so the number of rules has kind of
exploded for numbers.  There probably is a better way of handling
this, with better regex, but for now this is what I have.

I have also added a bunch of new unit tests for this.